+++ packages/react/src/icons/index.ts
@@ -6,6 +6,7 @@ export { IconFacebook } from './some/IconFacebook';
 export { IconGoogle } from './some/IconGoogle';
 export { IconInstagram } from './some/IconInstagram';
 export { IconLinkedin } from './some/IconLinkedin';
+export { IconRss } from './some/IconRss';
 export { IconSnapchat } from './some/IconSnapchat';
 export { IconTiktok } from './some/IconTiktok';
 export { IconTwitch } from './some/IconTwitch';
@@ -134,7 +135,6 @@ export { IconPodcast } from './ui/IconPodcast';
 export { IconPrinter } from './ui/IconPrinter';
 export { IconQuestionCircleFill } from './ui/IconQuestionCircleFill';
 export { IconRefresh } from './ui/IconRefresh';
-export { IconRss } from './ui/IconRss';
 export { IconSortAscending } from './ui/IconSortAscending';
 export { IconSortAlphabeticalAscending } from './ui/IconSortAlphabeticalAscending';
 export { IconSliders } from './ui/IconSliders';
